pyproject.toml
[tool.poetry]
name = "mlops-end-to-end"
version = "0.1.0"
description = ""
authors = ["picsalex <132259399+picsalex@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
minio = "^7.1.17"
python-decouple = "^3.8"
ulid-py = "^1.1.0"
aiohttp = "^3.8.6"
datasets = "^2.14.6"
pillow = "^10.1.0"
numpy = "^1.26.1"
s3fs = "^2023.10.0"
zenml = "^0.54.0"
ultralytics = "^8.0.232"
torch = "^2.1.2"
hydra-core = "^1.3.2"
hub-sdk = "^0.0.3"
mlflow = {extras = ["extra"], version = "^2.9.2"}


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"
ruff = "^0.1.5"
pylint = "^3.0.3"
mypy = "^1.8.0"
black = {extras = ["d"], version = "^23.12.1"}
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.isort]
known-local-folder = ["src"]
